feat: draw cursor in inverted colors
this is pretty cool. but what's cooler is how I got there! I changed
the definition on line. and updates 'btns.updater on the fly. then set
mark, new start and it.

this gives "in circuit serial programming" a whole new meaning! the
programmer is the device itself (and it has a compiler). all this work
finally starting to really pay off!